DO NOT MERGE : bad workaroud to remaning bug
This adds a sleep in renderReplacedElement so that drawimage works properly when the fix to svg files is used.
Without that, it looks like drawImage fires too fast and the fix is not yet done (code is async here)
So far I did not find out where I messed up with async code. Any help would be welcome !